+++ packages/nextjs-config/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @posthog/nextjs-config
 
+## 1.9.13
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @posthog/webpack-plugin@1.4.13
+
 ## 1.9.12
 
 ### Patch Changes